healthcare_analytics/pyproject.toml
[project]
name = "healthcare-analytics"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleksandr Balashenko",email = "balashenkoaleksandr294@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.132.0,<0.133.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "fastapi-pagination (>=0.15.10,<0.16.0)",
    "black (>=26.1.0,<27.0.0)"
]

[tool.black]
line-length = 120


[tool.ruff]
fix = false
lint.select = ["I001", "E402", "Q000", "ANN001", "ANN002", "ANN003", "ANN201", "ANN202", "ANN204", "ANN205", "ANN206", "F401", "F541", "E712", "W291", "E302", "A003", "A002", "A001"]
lint.ignore = ["ANN002", "ANN003", 'ANN204']
lint.fixable = []
lint.unfixable = []
line-length = 120
exclude = ["venv", ".venv", ".env"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
